chore: move login to components
